Delete User Feature Implemented
Delete user now works. Updates the database directly with u.delete() command, hence the changes to the db.sqlite3 file. There is a small error when returning to the main page after saying 'no' on confirm.html where it does not show the user's name. Will fix before Sprint 3 on 4/8